test: Merge test_ext_autodoc.py and test_autodoc.py
